now is't our problem

+++ Dragonet-core/src/main/java/org/dragonet/net/DragonetSession.java
@@ -214,7 +214,7 @@ public abstract class DragonetSession extends GlowSession {
         } else {
             pkStartGame.gamemode = 0;
         }
-        pkStartGame.eid = this.player.getEntityId();
+        pkStartGame.eid = (long)this.player.getEntityId();
         pkStartGame.spawnX = this.player.getWorld().getSpawnLocation().getBlockX();
         pkStartGame.spawnY = this.player.getWorld().getSpawnLocation().getBlockY();
         pkStartGame.spawnZ = this.player.getWorld().getSpawnLocation().getBlockZ();
